wip(poly check): refactor into a commands component, used by cli and poetry plugin. Bump dependencies.
