+++ components/polylith/distributions/__init__.py
@@ -1,0 +1,3 @@
+from polylith.distributions.core import distributions_packages
+
+__all__ = ["distributions_packages"]
