+++ components/polylith/poetry/internals.py
@@ -1,32 +1,9 @@
-from functools import reduce
 from pathlib import Path
-from typing import Dict, Iterable, List, Set, Union
+from typing import Iterable, Set, Union
 
 from poetry.factory import Factory
 from poetry.poetry import Poetry
 from poetry.utils.env import EnvManager
-
-
-def top_level_packages(dist) -> List[str]:
-    top_level = dist.read_text("top_level.txt")
-
-    return str.split(top_level or "")
-
-
-def mapped_packages(dist) -> dict:
-    packages = top_level_packages(dist)
-    name = dist.metadata["name"]
-
-    return {name: packages} if packages else {}
-
-
-def map_packages(acc, dist) -> dict:
-    return {**acc, **mapped_packages(dist)}
-
-
-def distributions_packages(dists) -> Dict[str, List[str]]:
-    """Return a mapping of top-level packages to their distributions."""
-    return reduce(map_packages, dists, {})
 
 
 def get_project_poetry(poetry: Poetry, path: Union[Path, None]) -> Poetry:
