+++ projects/poetry_polylith_plugin/pyproject.toml
@@ -38,12 +38,6 @@ poetry = "^1.2"
 tomlkit = "^0.11.5"
 rich = "^13.6.0"
 
-[tool.poetry.group.dev.dependencies]
-black = "^22.10.0"
-isort = "^5.10.1"
-mypy = "^0.982"
-flake8 = "^5.0.4"
-
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
