+++ pyproject.toml
@@ -31,19 +31,18 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.8.1"
 poetry = "^1.2"
 tomlkit = "^0.11.5"
 rich = "^13.6.0"
 typer = {extras = ["all"], version = "^0.9.0"}
 
 [tool.poetry.group.dev.dependencies]
-black = "^22.10.0"
+black = "^23.12.1"
 isort = "^5.10.1"
-mypy = "^0.982"
-flake8 = "^5.0.4"
-pytest = "^7.2.1"
-pytest-cov = "^4.0.0"
+mypy = "^1.8.0"
+flake8 = "^7.0.0"
+pytest = "^7.4.4"
 
 [tool.isort]
 profile = "black"
